requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
alembic==1.18.4
psycopg2-binary==2.9.12
pydantic==2.10.6
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.1.2
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
cryptography==44.0.3
